Data id type is changed from int to long

+++ src/main/java/godev/budgetgo/models/data/abstractions/DataBuilder.java
@@ -1,7 +1,7 @@
 package godev.budgetgo.models.data.abstractions;
 
 public interface DataBuilder<T extends Data> {
-    DataBuilder<T> setId(int id);
+    DataBuilder<T> setId(long id);
 
     T create();
 }
