+++ src/main/java/godev/budgetgo/models/repositories/abstractions/DataSpecification.java
@@ -1,5 +1,5 @@
 package godev.budgetgo.models.repositories.abstractions;
 
 public interface DataSpecification {
-    DataSpecification whereId(int id);
+    DataSpecification whereId(long id);
 }
